Task 3: Created separate files for common functions for reusability.
