+++ libs/stocks/feature-shell/src/lib/stocks/stocks.component.html
@@ -1,13 +1,9 @@
-<form class= "tp-form" [formGroup]="stockPickerForm">
+<form class="tp-form" [formGroup]="stockPickerForm">
   <mat-form-field class="tp-full-width">
     <mat-label>stocks symbol</mat-label>
-    <input type="text" placeholder="Symbol e.g AAPL" autocomplete="off"
-           aria-label="stocks symbol" matInput formControlName="symbol"
-      value=""/>
+    <input type="text" placeholder="Symbol e.g AAPL" autocomplete="off" aria-label="stocks symbol" matInput formControlName="symbol" value=""/>
     <mat-error><span
-        *ngIf="
-          !stockPickerForm.get('symbol').valid &&
-          stockPickerForm.get('symbol').touched">Please enter a symbol</span>
+      *ngIf="!stockPickerForm.get('symbol').valid && stockPickerForm.get('symbol').touched">Please enter a symbol</span>
     </mat-error>
   </mat-form-field>
 
@@ -20,8 +16,7 @@
 
   <mat-form-field>
     <mat-label>To Date </mat-label>
-    <input matInput [matDatepicker]="pickerToDate" [max]="max"
-           [min]="stockPickerForm.controls['fromDate'].value" formControlName="toDate">
+    <input matInput [matDatepicker]="pickerToDate" [max]="max" [min]="stockPickerForm.controls['fromDate'].value" formControlName="toDate">
     <mat-datepicker-toggle matSuffix [for]="pickerToDate"></mat-datepicker-toggle>
     <mat-datepicker #pickerToDate></mat-datepicker>
     <mat-error *ngIf="stockPickerForm.controls['toDate'].invalid">
@@ -32,4 +27,4 @@
   <button (click)="fetchQuote()" mat-button [disabled]="stockPickerForm.invalid" mat-raised-button>Go</button>
 </form>
 
-<coding-challenge-chart [data$]="quotes$"></coding-challenge-chart>
+<coding-challenge-chart [data$]="quotes$" [title]="'Stocks'" [type]="'LineChart'"></coding-challenge-chart>
