+++ libs/stocks/feature-shell/src/utils/stock-chart-constants.ts
@@ -2,6 +2,7 @@ export const CHART_DATE_RANGE = {
   PERIOD: {
     '1D': '1d',
     '5D': '5d',
+    '15D': '15d',
     '1M': '1m',
     '6M': '6m',
     '1Y': '1y',
